Allow returning errors when generating SQL

+++ integration/ref/ast/delete.go
@@ -19,6 +19,9 @@ func (s *DeleteStatement) GetQueryContext() *QueryContext {
 }
 
 func (s *DeleteStatement) toSQL(builder *strings.Builder, params *[]any, ctx *QueryContext) {
+	if ctx != nil && ctx.Error != nil {
+		return
+	}
 	s.context = ctx
 	builder.WriteString("DELETE FROM " + s.Table)
